# Add SnoopPrecompile directive